Fix import detail update quantity

+++ supplier-service/src/main/java/com/fromlabs/inventory/supplierservice/imports/details/ImportDetailEntity.java
@@ -36,7 +36,7 @@ public class ImportDetailEntity extends ImportReferencedMultiEntity {
 
     @Min(value = 0)
     @Column(name="quantity")
-    private Float quantity;
+    private Integer quantity;
 
     /**
      * Update entity by request
